refactor(slatetohtml): move conversion function to utilities
