+++ jest.config.js
@@ -1,9 +1,8 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
   verbose: true,
-  testEnvironment: 'node',
+  testEnvironment: 'jsdom',
   preset: 'ts-jest',
-  testEnvironment: 'node',
   testMatch: [
     '**/__tests__/**/*.spec.ts',
   ],
